Elasticsearch indexing warning fix
Made the Elasticsearch-indexing function skip the museum indexing if
there are no museum records in the relational database.

I made this change because running the script just to create the museum
index+mapping without data showed warinings in the elasticsearch logs
because sending a bulk index request without a body is not allowed.